Escape nested quote with an additional quote
According to RFC 4180 the proper way to escape double-quotes in CSV data
is to prepend an additional double-quote character:

> If double-quotes are used to enclose fields, then a double-quote
> appearing inside a field must be escaped by preceding it with
> another double quote.

Currently we escape double-quotes using a backslash. Opening such a file
in Microsoft Excel and/or Apple Numbers breaks the column layout, i.e.
the value containing the double quote is interpretted as multiple
columns instead of one. On the contrary, escaping using an extra
double-quote preserves the expected column layout in both Microsoft
Excel and Apple Numbers.

See also:
 - https://tools.ietf.org/html/rfc4180