Fix #149 (#150)
* Handle invalid parsed values from JSON.parse in csv2json

Since JSON.parse "interprets" some values, such as 39e1804 which is converted to Infinity. However, this isn't desirable since we want to return as close to the original value as possible. This commit adds some minor logic to verify that JSON.parse didn't determine the value is Infinity or -Infinity before setting the value in the returned JSON. If this is the case, the value is simply returned as the same string that's contained in the CSV.

Fixes #149

* Version bump to 3.7.1