Handle empty array case when unwindArrays option is true.
If an empty array was encountered when the unwindArrays option was set to true, the document was ignored and essentially was removed from the CSV output which is not the expected behavior. This commit addresses that bug by ensuring that the "unwound" version of the empty array is still pushed to the accumulator so the document is not excluded.

Fixes #151.