Fix to properly handle empty last field values in csv2json.
CSV2JSON was not properly handling empty field values that occurred at the end of the CSV. This resulted in a missing field in the last JSON document and the prior field value containing a comma at the end. An additional branch was added to address this missing case.

Fixes #155