Add an option to specify a custom value parser fn.
One of the commonly requested features for both json2csv and csv2json is
the ability to specify a custom value parser function. This
functionality also provides a way to customize the behavior of the
library as well to ensure that custom values are converted correctly
when converting to the other format.