Validate field titles against Object.keys list (#186)
* Validate field titles against Object.keys list

As reported in #185, there was a bug where header values could not have
the same name as a Map method. This is because the code was naively
checking for the existence of a possible title field by simply accessing
the raw value of the Map. This resulted in Map.prototype.set being
returned when the "set" JSON key name was used. The logic has been
updated to check for header titles against an Object.keys list now since
that will not include prototype methods for the underlying Map object.

Fixes #185

* Add test for #185.

* Add Node 16,  remove Node 10 for travis test cfg.