Support keys with '.' characters in them (#194)
* Add support for keys with nested dots.

Through modifications to `doc-path` and `deeks` dependency underlying
code, I've worked to add support for the keys with nested dots in the
JSON key path and conversion back to CSV. This was achieved through the
`deeks` module's new `escapeNestedDots` option which will now be set to
`true`. Additionally, support for this escaping was added to `doc-path`
so that the correct values are retrieved and can also be correctly set
for csv2json. This commit adds the option specification for the `deeks`
module and the updated versions of the dependency modules in
package.json.

Fixes #184