Update doc-path to fix negative lookbehind regexp. (#199)
* Update doc-path to fix negative lookbehind regexp.

As reported in #197, there was a compatibility issue introduced in
doc-path@3.0.0 where a negative lookbehind Regexp was being used despite
not being supported by many browsers. This caused runtime errors when
this library's functionality was called due to the underlying dependency
compatibility issue. This commit bumps the version to 3.0.1 which fixes
the compatibility issue in the doc-path module.

Fixes #197

* chore(release): 3.14.1