Fix 203 (#204)
* Handle case where no column headers are present.

As reported in #203, there are use cases where the CSV may not have
column headers for a particular set of data. In this case, the expected
output simply ignores these unnamed data values, but as of doc-path@2.0.0,
the output was an error indicating that no key path was provided. This
commit fixes #203 by capturing and handling that error from the doc-path
module.

Fixes #203

* chore(release): 3.14.2