Handle nested newlines in quoted fields.
As reported in #214, there was a bug with how the CSV parser was handling quoted fields which had newlines (or EOL delimiters) embedded in them. Since the value is wrapped in quotes, the EOL delimiter should not be interpreted as though it's a new line or the start of another field as it previously was. This commit introduces a fix to ensure that quoted fields are parsed properly regardless of whether there is a nested EOL delimiter inside.

Fixes #214